+++ packages/calamares-config-xfce/files/post_install_nvidia.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-## Copyright (C) 2020-2023 Aditya Shakya <adi1090x@gmail.com>
+## Copyright (C) 2020-2024 Aditya Shakya <adi1090x@gmail.com>
 ## Everyone is permitted to copy and distribute copies of this file under GNU-GPL3
 
 ## Post installation script for Archcraft (only to detect nvidia)